refactor multiPublishingBeaconNodeApi
split out the ValidatorApiChannel interface.

partially addresses #4223
